pyproject.toml
[project]
name = "transcribe-audio"
version = "0.1.0"
description = "AI-powered audio transcription with intelligent summarization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.74.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "W", "I", "N", "B", "C4", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
